[common][controller][protocol] Changes to allow metadata store shared Zk creation and materialize metadata store version for a Venice store
1. Three new Admin endpoint to create new versions for Zk shared stores and materialize/dematerialize a metadata store version for a given Venice store:
     - newZkSharedStoreVersion
     - materializeMetadataStoreVersion
     - dematerializeMetadataStoreVersion

2. New Store field storeMetadataSystemStoreEnabled to indicate whether a metadata store is materialized for the Venice store or not.

3. New controller client methods to create zk shared stores and create them with default system store configs (hybrid enabled).

4. The correct order to set up metadata store for a given Venice store should be:
     - Create the Zk shared store for the metadata store type with name METADATA_STORE.getPrefix()
     - Create a new version for the Zk shared metadata store object using newZkSharedStoreVersion
     - Materialize the metadata store version using materializeMetadataStoreVersion

5. Metadata store is enabled/disabled at a store level because this way we minimize the impact to regular stores. In the future we can have metadata
store version created and configured by default like some of the other system stores.

RB=2108738
BUG=VENG-5783
G=voldemort-reviewers
R=zpolicze
A=zpolicze

+++ venice-controller/src/main/java/com/linkedin/venice/controller/kafka/consumer/AdminExecutionTask.java
@@ -1,5 +1,6 @@
 package com.linkedin.venice.controller.kafka.consumer;
 
+import com.linkedin.venice.common.VeniceSystemStore;
 import com.linkedin.venice.common.VeniceSystemStoreUtils;
 import com.linkedin.venice.compression.CompressionStrategy;
 import com.linkedin.venice.controller.ExecutionIdAccessor;
@@ -321,7 +322,13 @@ public class AdminExecutionTask implements Callable<Void> {
     String clusterName = message.clusterName.toString();
     String storeName = message.storeName.toString();
     int versionNum = message.versionNum;
-    admin.deleteOldVersionInStore(clusterName, storeName, versionNum);
+    if (VeniceSystemStoreUtils.getSystemStore(storeName) == VeniceSystemStore.METADATA_STORE) {
+      // Dematerialize a metadata store version.
+      admin.dematerializeMetadataStoreVersion(clusterName, VeniceSystemStoreUtils.getStoreNameFromMetadataStoreName(storeName), versionNum);
+    } else {
+      // Delete an old version for a Venice store.
+      admin.deleteOldVersionInStore(clusterName, storeName, versionNum);
+    }
     logger.info("Deleted version: " + versionNum + " in store:" + storeName + " in cluster: " + clusterName);
   }
 
@@ -509,8 +516,18 @@ public class AdminExecutionTask implements Callable<Void> {
             pushType, remoteKafkaBootstrapServers);
       }
     } else {
-      admin.addVersionAndStartIngestion(clusterName, storeName, pushJobId, versionNumber, numberOfPartitions, pushType,
-          remoteKafkaBootstrapServers);
+      if (VeniceSystemStore.METADATA_STORE.getPrefix().equals(storeName)) {
+        // New version for the Zk shared metadata store.
+        admin.newZkSharedStoreVersion(clusterName, storeName);
+      } else if (VeniceSystemStoreUtils.getSystemStore(storeName) == VeniceSystemStore.METADATA_STORE) {
+        // Materialize a metadata store for a specific Venice store.
+        admin.materializeMetadataStoreVersion(clusterName,
+            VeniceSystemStoreUtils.getStoreNameFromMetadataStoreName(storeName), versionNumber);
+      } else {
+        // New version for regular Venice store.
+        admin.addVersionAndStartIngestion(clusterName, storeName, pushJobId, versionNumber, numberOfPartitions, pushType,
+            remoteKafkaBootstrapServers);
+      }
     }
   }
